Added separate work item link design file for clarity
